+++ src/components/practice/CategoriesList.jsx
