+++ src/components/practice/CategoryItem.jsx
