+++ src/components/practice/CategoryItem.test.jsx
