+++ src/components/practice/InterviewCategories.jsx
