+++ src/pages/interviews/InterviewsPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import InterviewsContainer from '../../containers/InterviewsContainer';
+import InterviewsContainer from '../../containers/interviews/InterviewsContainer';
 import BaseLayout from '../../layout/BaseLayout';
 
 const InterviewsPage = () => (
